requirements.txt
certifi==2018.4.16
chardet==3.0.4
docker==3.4.1
docker-pycreds==0.3.0
idna==2.7
pypiwin32==220
requests==2.19.1
six==1.11.0
#tornado==5.1
urllib3==1.23
websocket-client==0.48.0
wsrpc-tornado==0.5.6
